Emit assembly attributes after Import statements
The VB language requires that code files have statements in the
following order:

- Option
- Import
- Assembly level attributes
- Namespaces / other declarations

The VBCodeGenerator type when given both import and assembly level
attribute statements was generating the attributes first.  This lead to
compilation errors in the generated file.

This change fixes the generator to emit assembly level attributes in
the proper location in the file.